[#214] feat: group list UI 구현 및 파일 분리
