+++ lib/common/constants/app_colors.dart
@@ -32,6 +32,7 @@ class CustomColors {
   static const Color whRedBright = Color(0xffDA3A3A);
   static const Color whUnSelectedTextColor = Color(0xffA2A2A2);
   static const Color whSelectedTextColor = Color(0xff000000);
+  static const Color whPlaceholderGrey = Color(0xffe3e3e3);
 }
 
 class PointColors {
