+++ lib/presentation/group/view/view.dart
@@ -1,1 +1,2 @@
+export 'group_list_view_cell_widget.dart';
 export 'group_view.dart';
